pyproject.toml
[project]
name = "newapi-credit-logger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "apscheduler>=3.11.2",
    "cryptography>=46.0.5",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "pymysql>=1.1.2",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
